backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
python-decouple==3.8
dj-database-url==2.2.0
psycopg2-binary==2.9.9
redis==5.2.0
django-redis==5.4.0
celery==5.4.0
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
sentry-sdk==2.19.0